Optimize ovsconfig dependency
1. Remove package dependent of pkg/agent/util from ovs_client_windows.
2. Update MTU on both OVS port and host interface on Windows when
   setting MTU on antrea-gw0.
