+++ pkg/ovs/ovsconfig/ovs_client.go
@@ -970,3 +970,23 @@ func (br *OVSBridge) SetPortExternalIDs(portName string, externalIDs map[string]
 	}
 	return nil
 }
+
+func (br *OVSBridge) SetInterfaceMTU(name string, MTU int) error {
+	tx := br.ovsdb.Transaction(openvSwitchSchema)
+
+	tx.Update(dbtransaction.Update{
+		Table: "Interface",
+		Where: [][]interface{}{{"name", "==", name}},
+		Row: map[string]interface{}{
+			"mtu_request": MTU,
+		},
+	})
+
+	_, err, temporary := tx.Commit()
+	if err != nil {
+		klog.Error("Transaction failed: ", err)
+		return NewTransactionError(err, temporary)
+	}
+
+	return nil
+}
